api/requirements.txt
alembic==1.2.1
psycopg2-binary==2.8.6
aniso8601==9.0.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
click==7.0
cryptography==3.2
entrypoints==0.3
flask-cors==3.0.10
flask-migrate==2.5.2
flask-restful==0.3.8
flask-sqlalchemy==2.4.1
flask-wtf==0.14.2
flask==1.1.1
idna==2.8
itsdangerous==1.1.0
jeepney==0.4.1
jinja2==2.10.3
jsonconfig-tool==1.5
keyring==19.2.0
mako==1.1.0
markupsafe==1.1.1
messages==0.5.0
passlib==1.7.4
pycparser==2.19
pyjwt==2.0.1
python-box==3.4.5
python-dateutil==2.8.0
python-decouple==3.1
python-editor==1.0.4
pytz==2021.1
requests==2.22.0
secretstorage==3.1.1; sys_platform == 'linux'
six==1.12.0
sqlalchemy==1.3.10
urllib3==1.25.6
validus==0.3.0
werkzeug==0.16.0
wtforms==2.2.1